Convert series tests to mocha
