+++ test/test-async.js
@@ -16,34 +16,10 @@ if (!Function.prototype.bind) {
     };
 }
 
-function getFunctionsObject(call_order) {
-    return {
-        one: function(callback){
-            setTimeout(function(){
-                call_order.push(1);
-                callback(null, 1);
-            }, 125);
-        },
-        two: function(callback){
-            setTimeout(function(){
-                call_order.push(2);
-                callback(null, 2);
-            }, 200);
-        },
-        three: function(callback){
-            setTimeout(function(){
-                call_order.push(3);
-                callback(null, 3,3);
-            }, 50);
-        }
-    };
-}
-
 function isBrowser() {
     return (typeof process === "undefined") ||
         (process + "" !== "[object process]"); // browserify
 }
-
 
 exports['seq'] = function (test) {
     test.expect(5);
@@ -151,206 +127,6 @@ exports['seq without callback'] = function (test) {
     var add2mul3 = async.seq(add2, mul3);
     add2mul3.call(testcontext, 3);
 };
-
-exports['series'] = {
-
-    'series': function(test){
-    var call_order = [];
-    async.series([
-        function(callback){
-            setTimeout(function(){
-                call_order.push(1);
-                callback(null, 1);
-            }, 25);
-        },
-        function(callback){
-            setTimeout(function(){
-                call_order.push(2);
-                callback(null, 2);
-            }, 50);
-        },
-        function(callback){
-            setTimeout(function(){
-                call_order.push(3);
-                callback(null, 3,3);
-            }, 15);
-        }
-    ],
-    function(err, results){
-        test.ok(err === null, err + " passed instead of 'null'");
-        test.same(results, [1,2,[3,3]]);
-        test.same(call_order, [1,2,3]);
-        test.done();
-    });
-},
-
-    'with reflect': function(test){
-    var call_order = [];
-    async.series([
-        async.reflect(function(callback){
-            setTimeout(function(){
-                call_order.push(1);
-                callback(null, 1);
-            }, 25);
-        }),
-        async.reflect(function(callback){
-            setTimeout(function(){
-                call_order.push(2);
-                callback(null, 2);
-            }, 50);
-        }),
-        async.reflect(function(callback){
-            setTimeout(function(){
-                call_order.push(3);
-                callback(null, 3,3);
-            }, 15);
-        })
-    ],
-    function(err, results){
-        test.ok(err === null, err + " passed instead of 'null'");
-        test.deepEqual(results, [
-            { value: 1 },
-            { value: 2 },
-            { value: [3,3] }
-        ]);
-        test.same(call_order, [1,2,3]);
-        test.done();
-    });
-},
-
-    'empty array': function(test){
-    async.series([], function(err, results){
-        test.equals(err, null);
-        test.same(results, []);
-        test.done();
-    });
-},
-
-    'error': function(test){
-    test.expect(1);
-    async.series([
-        function(callback){
-            callback('error', 1);
-        },
-        function(callback){
-            test.ok(false, 'should not be called');
-            callback('error2', 2);
-        }
-    ],
-    function(err){
-        test.equals(err, 'error');
-    });
-    setTimeout(test.done, 100);
-},
-
-    'error with reflect': function(test){
-    test.expect(2);
-    async.series([
-        async.reflect(function(callback){
-            callback('error', 1);
-        }),
-        async.reflect(function(callback){
-            callback('error2', 2);
-        }),
-        async.reflect(function(callback){
-            callback(null, 1);
-        })
-    ],
-    function(err, results){
-        test.ok(err === null, err + " passed instead of 'null'");
-        test.deepEqual(results, [
-            { error: 'error' },
-            { error: 'error2' },
-            { value: 1 }
-        ]);
-        test.done();
-    });
-},
-
-    'no callback': function(test){
-    async.series([
-        function(callback){callback();},
-        function(callback){callback(); test.done();},
-    ]);
-},
-
-    'object': function(test){
-    var call_order = [];
-    async.series(getFunctionsObject(call_order), function(err, results){
-        test.equals(err, null);
-        test.same(results, {
-            one: 1,
-            two: 2,
-            three: [3,3]
-        });
-        test.same(call_order, [1,2,3]);
-        test.done();
-    });
-},
-
-    'call in another context': function(test) {
-    if (isBrowser()) {
-        // node only test
-        test.done();
-        return;
-    }
-    var vm = require('vm');
-    var sandbox = {
-        async: async,
-        test: test
-    };
-
-    var fn = "(" + (function () {
-        async.series([function (callback) {
-            callback();
-        }], function (err) {
-            if (err) {
-                return test.done(err);
-            }
-            test.done();
-        });
-    }).toString() + "())";
-
-    vm.runInNewContext(fn, sandbox);
-},
-
-    // Issue 10 on github: https://github.com/caolan/async/issues#issue/10
-    'falsy return values': function (test) {
-    function taskFalse(callback) {
-        async.nextTick(function() {
-            callback(null, false);
-        });
-    }
-    function taskUndefined(callback) {
-        async.nextTick(function() {
-            callback(null, undefined);
-        });
-    }
-    function taskEmpty(callback) {
-        async.nextTick(function() {
-            callback(null);
-        });
-    }
-    function taskNull(callback) {
-        async.nextTick(function() {
-            callback(null, null);
-        });
-    }
-    async.series(
-        [taskFalse, taskUndefined, taskEmpty, taskNull],
-        function(err, results) {
-            test.equal(results.length, 4);
-            test.strictEqual(results[0], false);
-            test.strictEqual(results[1], undefined);
-            test.strictEqual(results[2], undefined);
-            test.strictEqual(results[3], null);
-            test.done();
-        }
-    );
-}
-
-};
-
 
 exports['iterator'] = function(test){
     var call_order = [];
